+++ src/components/message/error-message/error-message.tsx
@@ -1,0 +1,12 @@
+import React, { HTMLAttributes } from "react";
+import styles from "./error-message.module.scss";
+
+type Props = {} & HTMLAttributes<HTMLDivElement>;
+
+export const ErrorMessage = ({ children, ...props }: Props) => {
+  return (
+    <p className={styles.text} {...props}>
+      {children}
+    </p>
+  );
+};
