+++ src/components/message/error-message/index.ts
@@ -1,0 +1,1 @@
+export * from "./error-message";
