+++ src/components/message/success-message/index.ts
@@ -1,0 +1,1 @@
+export * from "./success-message";
