+++ src/components/message/success-message/success-message.module.scss
@@ -1,0 +1,9 @@
+.text {
+  color: $success-accent-color;
+  margin-bottom: 28px;
+
+  &::before {
+    content: "\2022";
+    margin-right: 10px;
+  }
+}
