+++ src/components/modals/modal-auth/modal-auth.module.scss
@@ -6,90 +6,33 @@
   position: relative;
   box-shadow: 1px 1px 15px rgba(0, 0, 0, 0.125);
   box-sizing: border-box;
-  min-height: 720px;
+  min-height: 738px;
   max-height: 100vh;
 }
 
-.backgroundSignIn {
-  @include background;
-  position: absolute;
-  background-position: 50%;
-  background-repeat: no-repeat;
-  position: absolute;
-  top: 34px;
-  left: 76px;
-}
-
-.backgroundResetPassword {
-  position: absolute;
-  top: 73px;
-  left: 76px;
-}
-
-.form {
-  position: relative;
-}
-
-.inputsRegistration {
-  display: grid;
-  grid-template-columns: [startline] repeat(2, minmax(296px, max-content)) [endline];
-  column-gap: 44px;
-  row-gap: 28px;
-  margin-bottom: 36px;
-
-  @media screen and (max-width: 1024px) {
-    grid-template-columns: 1fr;
-  }
-}
-
-.inputsResetPassword {
-  display: grid;
-  margin-bottom: 28px;
-  max-width: 296px;
-  gap: 20px;
-}
-
-.inputsSignin {
-  display: grid;
-  row-gap: 20px;
-  margin-bottom: 8px;
-  max-width: 296px;
-}
-
-.wrapperField {
-  max-width: "296px";
-  width: 100%;
-}
-
-.listButtons {
-  list-style: none;
-  display: flex;
-}
-
 .containerButtons {
+  display: flex;
   justify-content: space-between;
   border-radius: 18px 18px 0 0;
   background-color: $almond;
 }
 
-.itemLink {
+.item {
   width: 100%;
 }
 
-.itemButtons {
-  margin-right: 28px;
-
-  &:last-of-type {
-    margin-right: 0;
-  }
-}
-
 .buttonLink {
-  @include reset-button;
+  border: none;
+  cursor: pointer;
   height: 72px;
   width: 100%;
   background-color: $almond;
   border-radius: 18px 18px 0 0;
+}
+
+.activeNav {
+  background-color: $white !important;
+  box-shadow: 0px -8px 10px -6px rgba(0, 0, 0, 0.09);
 }
 
 .activeButtonText {
@@ -107,129 +50,8 @@
   }
 }
 
-.activeNav {
-  background-color: $white !important;
-  box-shadow: 0px -8px 10px -6px rgba(0, 0, 0, 0.09);
-}
-
 .content {
-  padding: 50px 0 50px 75px;
   position: relative;
   z-index: 1;
   background-color: white;
-
-  @media screen and (max-width: 1024px) {
-    padding: 48px 24px;
-  }
-}
-
-.title {
-  @include headline;
-  @include headline2;
-
-  position: relative;
-  margin-bottom: 36px;
-  max-width: 458px;
-}
-
-.titleSubtext {
-  margin-bottom: 10px;
-}
-
-.label {
-  display: flex;
-  flex-direction: column;
-  cursor: text;
-  max-width: 296px;
-}
-
-.resetPasswordLabel {
-  grid-column: startline/endline;
-
-  &:last-of-type {
-    margin-bottom: 36px;
-  }
-}
-
-.positionCenter {
-  display: block;
-  margin: 0 auto;
-}
-
-.subtext {
-  max-width: 458px;
-  margin-bottom: 28px;
-}
-
-.textField {
-  max-width: 296px;
-  border-radius: 10px;
-  border: 1px solid $light-gray;
-  padding: 12px 16px;
-
-  &::placeholder {
-    @include text;
-    @include textSmall;
-
-    color: $asphalt;
-  }
-}
-
-.errorsText {
-  grid-column: startline/endline;
-  padding: 0 20px;
-  list-style: none;
-  margin-bottom: 36px;
-  display: grid;
-  row-gap: 16px;
-}
-
-.textError {
-  margin-bottom: 28px;
-  &::before {
-    content: "\2022";
-    margin-right: 10px;
-  }
-}
-
-.textSuccess {
-  color: $success-accent-color;
-  margin-bottom: 28px;
-  &::before {
-    content: "\2022";
-    margin-right: 10px;
-  }
-}
-
-.warningText {
-  color: $pink-pig;
-}
-
-.checkboxLabel {
-  display: flex;
-  align-items: center;
-  grid-column: startline/endline;
-  margin-bottom: 36px;
-  cursor: pointer;
-}
-
-.linkText {
-  @include reset-button;
-
-  background-color: inherit;
-  cursor: pointer;
-  margin: 4px 0 46px;
-  color: $black;
-  display: flex;
-  padding-left: 16px;
-}
-
-.illustration {
-  position: absolute;
-  top: 96px;
-  right: 32px;
-}
-
-.checkbox {
-  margin-bottom: 52px;
 }
