+++ src/components/modals/modal-auth/registration/index.ts
