+++ src/components/modals/modal-auth/registration/registration.module.scss
@@ -1,0 +1,35 @@
+.container {
+  margin: 50px 92px 64px 76px;
+
+  @media screen and (max-width: 1024px) {
+    margin: 48px 24px;
+  }
+}
+
+.containerTitle {
+  margin-bottom: 28px;
+}
+
+.title {
+  margin-bottom: 12px;
+}
+
+.containerForm {
+  display: grid;
+  grid-template-columns: [startline] repeat(2, minmax(296px, max-content)) [endline];
+  column-gap: 44px;
+  row-gap: 28px;
+  margin-bottom: 36px;
+
+  @media screen and (max-width: 1024px) {
+    grid-template-columns: 1fr;
+  }
+}
+
+.checkbox {
+  margin-bottom: 52px;
+}
+
+.positionCenter {
+  margin: 0 auto;
+}
