+++ src/components/modals/modal-auth/reset-password-confirm/index.ts
