+++ src/components/modals/modal-auth/reset-password-confirm/reset-password-confirm.module.scss
@@ -1,0 +1,34 @@
+.container {
+  padding: 86px 76px;
+}
+
+.content {
+  max-width: 458px;
+  position: relative;
+}
+
+.background {
+  position: absolute;
+  top: 73px;
+  left: 76px;
+}
+
+.containerTitle {
+  margin-bottom: 28px;
+}
+
+.title {
+  margin-bottom: 12px;
+}
+
+.containerInputs {
+  display: grid;
+  margin-bottom: 28px;
+  max-width: 296px;
+  row-gap: 20px;
+}
+
+.containerButtons {
+  display: flex;
+  column-gap: 20px;
+}
