+++ src/components/modals/modal-auth/reset-password/reset-password.module.scss
@@ -1,0 +1,33 @@
+.container {
+  padding: 86px 76px;
+}
+
+.content {
+  max-width: 458px;
+  position: relative;
+}
+
+.background {
+  position: absolute;
+  top: 73px;
+  left: 76px;
+}
+
+.containerTitle {
+  margin-bottom: 28px;
+}
+
+.title {
+  margin-bottom: 12px;
+}
+
+.containerInput {
+  width: 100%;
+  margin-bottom: 28px;
+  max-width: 296px;
+}
+
+.containerButtons {
+  display: flex;
+  column-gap: 20px;
+}
