+++ src/components/modals/modal/modal.module.scss
@@ -24,7 +24,7 @@
   justify-content: center;
 
   width: 100%;
-  max-width: 860px;
+  max-width: 864px;
   padding: 30px;
   border-radius: 4px;
   animation: show 150ms ease-in;
