+++ src/shared/ui/inputs/checkbox/checkbox.module.scss
@@ -61,25 +61,12 @@
   display: none;
 }
 
-.label {
-  @include textSmall;
-  @include text;
-
-  color: $black;
-}
-
 .warning {
   color: $pink-pig;
 }
 
-.errorTextMargins {
+.errorMessage {
   margin: 8px 0 0 44px;
-}
-
-.message {
-  @include textTiny;
-  @include text;
-
   display: inline-block;
   margin-top: 4px;
   color: $error-accent-color;
