+++ src/shared/ui/inputs/input/input.tsx
@@ -18,7 +18,7 @@ export const Input = forwardRef<HTMLInputElement, InputProps>(
     return (
       <div className={cn(styles.container, className)}>
         {label && (
-          <label className={cx("text", "label", { warning: error && !disabled })} htmlFor={id}>
+          <label className={cx("text-medium", "label", { labelError: error && !disabled })} htmlFor={id}>
             {error && <Icon glyph="warning" className={cx("warningIcon")} />}
 
             {label}
@@ -26,13 +26,14 @@ export const Input = forwardRef<HTMLInputElement, InputProps>(
         )}
 
         <input
+          id={id}
           ref={ref}
           disabled={disabled}
-          className={cx("input", "text", { disabled }, { error }, className)}
+          className={cx("input", "text-medium", { disabled }, { errorBorder: error }, className)}
           {...props}
         />
 
-        {errorMessage && error && <span className={cx("message")}>{errorMessage}</span>}
+        {errorMessage && error && <span className={cx("text-small", "errorMessage")}>{errorMessage}</span>}
       </div>
     );
   }
