+++ src/shared/ui/inputs/password/password.module.scss
@@ -5,8 +5,8 @@
 }
 
 .buttonShowPassword {
-  @include reset-button;
-
+  border: none;
+  cursor: pointer;
   position: absolute;
   right: 16px;
   top: 42px;
