+++ src/styles/global.css
@@ -123,6 +123,11 @@ h4 {
   justify-content: space-between;
 }
 
+.headline2 {
+  font-size: 30px;
+  line-height: 130%;
+}
+
 .subtitle1 {
   font-family: var(--font-open-sans), Arial, sans-serif;
   font-size: 20px;
